Add files via upload
Task description:
You need to design and develop the sleep function of the OS.




You should implement the function sleep(long timeInMilis). 
When the OS calls this function with a job, it is added to a dedicated sleeping jobs queue (you need to design and implement it as well). 
At the right time, the OS should bring the job back into the running jobs queue from which the OS will take and execute it (you don't need to implement that part)

You should also implement the function awake(). This function is called by the OS every 1ms and is responsible to move the sleeping jobs that are ready to execute to the running queue.

1. When you think about the design, think about performance issues. Which function should be more efficient and why?
2. What data structure will you use in order to keep the sleeping jobs queue?
3. The OS does not have a clock function. You don't have access to the current time.
4. You cannot use a global static variable to simulate current time